requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.4,<6
drf-spectacular>=0.29,<1
psycopg2-binary>=2.9,<3
django-environ>=0.11,<1
